backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.0
flask-migrate>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
nltk>=3.8.0
spacy>=3.6.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
celery>=5.3.0
redis>=4.6.0
arxiv>=1.4.0
feedparser>=6.0.0
schedule>=1.2.0
jupyter>=1.0.0
pytest>=7.4.0
pytest-flask>=1.2.0